refactor: move Swift cache policy to recording ownership
