+++ src/recording/overlay.ts
@@ -3,7 +3,7 @@ import path from 'node:path';
 import { fileURLToPath } from 'node:url';
 import { runCmd } from '@agent-device/host-kit/command';
 import { AppError } from '@agent-device/kernel/errors';
-import { buildSwiftToolEnv, compileSwiftSourceFile } from '../utils/swift-cache.ts';
+import { buildSwiftToolEnv, compileSwiftSourceFile } from './swift-cache.ts';
 import { findProjectRoot } from '@agent-device/host-kit/version';
 import { waitForPlayableVideo, waitForStableFile } from '../utils/video.ts';
 import {
